requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
brotli==1.0.9
    # via whitenoise
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.1.2
    # via -r requirements/requirements.in
certifi==2022.5.18.1
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==37.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.7.2
    # via -r requirements/requirements.in
django==3.2.16
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-filter
    #   django-hijack
    #   django-modelcluster
    #   django-otp
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   drf-spectacular
    #   jsonfield
    #   wagtail
django-allauth==0.50.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.6
    # via -r requirements/requirements.in
django-environ==0.9.0
    # via -r requirements/requirements.in
django-hijack==3.2.1
    # via -r requirements/requirements.in
django-otp==1.1.3
    # via django-allauth-2fa
django-waffle==3.0.0
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
    #   wagtail
djangorestframework-api-key==2.2.0
    # via -r requirements/requirements.in
drf-spectacular==0.22.1
    # via -r requirements/requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonfield==3.1.0
    # via dj-stripe
jsonschema==4.6.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via redis
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0
    # via drf-spectacular
qrcode==7.3.1
    # via django-allauth-2fa
redis==4.3.1
    # via celery
requests==2.27.1
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.5.12
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   html5lib
    #   l18n
sqlparse==0.4.2
    # via django
stripe==3.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise[brotli]==6.2.0
    # via -r requirements/requirements.in
wrapt==1.14.1
    # via deprecated
python-docx==0.8.11
    # via -r requirements/requirements.in